+++ src/main/java/com/salesforce/phoenix/expression/DecimalMultiplyExpression.java
@@ -54,13 +54,16 @@ public class DecimalMultiplyExpression extends MultiplyExpression {
         BigDecimal result = null;
         for (int i=0; i<children.size(); i++) {
             Expression childExpr = children.get(i);
-            if (!childExpr.evaluate(tuple, ptr) || ptr.getLength() == 0) { 
+            if (!childExpr.evaluate(tuple, ptr)) {
                 return false;
             }
-
+            if (ptr.getLength() == 0) {
+                return true;
+            }
+            
             PDataType childType = children.get(i).getDataType();
             BigDecimal bd= (BigDecimal)PDataType.DECIMAL.toObject(ptr, childType);
-
+            
             if (result == null) {
                 result = bd;
                 maxLength = childExpr.getMaxLength();
