+++ src/main/java/com/salesforce/phoenix/schema/PColumnImpl.java
@@ -157,8 +157,8 @@ public class PColumnImpl implements PColumn {
         Bytes.writeByteArray(output, familyName == null ? ByteUtil.EMPTY_BYTE_ARRAY : familyName.getBytes());
         WritableUtils.writeVInt(output, dataType.ordinal());
         // Use Integer.MIN_VALUE as a token for maxLength being null.
-        WritableUtils.writeVInt(output, maxLength == null ? Integer.MIN_VALUE : maxLength);
-        WritableUtils.writeVInt(output, scale == null ? Integer.MIN_VALUE : scale);
+        WritableUtils.writeVInt(output, maxLength == null ? QueryConstants.NO_MAXLENGTH : maxLength);
+        WritableUtils.writeVInt(output, scale == null ? QueryConstants.NO_SCALE : scale);
         output.writeBoolean(nullable);
         WritableUtils.writeVInt(output, position);
     }
